blanket migration for degree and trace added Eq support to operation few bugfixes
